'errors.DataConflictError' for duplicate iteration creation causes 500 error in API (OSIO#2865)
add support for `409 Conflict` response at the design level
also: renamed and refactored the test file, and include a verification for this bug fix.

Fixes openshiftio/openshift.io#2865
